test: split PDF bibliography classification cases
